requirements.txt
python-dotenv==1.0.0
selenium==4.27.1
webdriver-manager==4.0.2
requests==2.31.0
urllib3==2.0.7
pytz==2023.3
pytest==8.3.3
pytest-mock==3.14.0
psutil==5.9.8